Revert "Don't try (and fail) to refresh features' IsActive when unlocking"
This reverts commit 136a703027894f782d7afbefe7b7f5ed95481fa8, which was
a workaround for one case where the bad handling of inconsistent state in
the refresh_is_active method lead to a bug.

Commit "tps: Fix activation of feature with inconsistent state" is a
more general solution, which makes the workaround unnecessary:
refresh_is_active will not set IsActive for any features whose bindings
are not active, so it's a no-op after a call to Unlock (unless a symlink
feature is enabled in the config file and there are still valid
symlinks, in which case it's correct that the feature is treated as
active).