Additional Software: Handle package changes in systemd service
... instead of a child created via double fork. One problem with the
double fork is that the child fails wit an I/O error when it tries
to print to stderr after the parent has exited.

Note that this required creating the temporary file used to atomically
update live-additional-software.conf as root (and then changing the
owner) to avoid the operation to fail with a Permission Denied error.

Closes #19552