Make code easier to review
When reading this, I wondered if a GObject.Object could be False-ish. In fact, that's how if not self.switch works. So I needed to wonder, and guessed that it should never be the case, and then I can be sure I'm understanding what this is checking.
A more specific expression would have saved me this reasoning.

I don't think this is serious enough to deserve being changed in all our codebase, though.