Test suite: Print exception with stack trace
Instead of handling and printing exceptions, define a function and call
that to let any exception abort execution automatically and let the
default handler print the exception with a stack trace.

Also, define the script as a multiline string which allows us to copy
and paste it as is into a Python interpreter in a running Tails.