Test suite: Support configuring chutney start time
It sometimes happens on my local system that chutney doesn't start at
all and I have to delete /tmp/TailsToaster/chutney-data/ to fix that.
Currently I only notice that that's the case after the start time of 10
minutes has passed. Chutney never takes that long to start on my system,
so I want to reduce that startup time for my local runs.