Fix race when starting tails-ibus-proxy.service
I've seen the service fail during startup with:

    Traceback (most recent call last):
      File "/usr/local/lib/tails-ibus-proxy", line 22, in <module>
        main()
      File "/usr/local/lib/tails-ibus-proxy", line 16, in main
        address, guid = output.split(",")
    ValueError: not enough values to unpack (expected 2, got 1)

That caused TCA to not start because the service is a dependency.

The reason is probably that the ibus address is only known after the
ibus-daemon was started by org.gnome.Shell@wayland.service, so we're
now starting the service after that.