Test suite: Reduce number of snapshots
In the previous commit I created some snapshots which are currently not
used or only used once. Creating snapshots has a cost because it takes
time and they use up RAM (at least on Lizard where they are still stored
on a tmpfs), so we should not create snapshots unnecessarily.