XXX: Test suite: Use external snapshots
Handling external snapshots in the disk_path method, required moving
some of the functionality from the VMStorage class to the VM class,
because we need information from the domain to figure out the correct
disk image when using external snapshots.

In cases where we actually want a volume of the specified name,
volume_path() is now used.

Contains XXX debug output