Bump APT snapshots for the Vagrant box
Our apt-snapshots-expiry CI check is a bit simplistic: it assumes that the
Vagrant box uses the same version of Debian as the Tails image we're building,
so on the feature/bookworm branch it checks snapshots of the bookworm-security
suite, which we did not create in December 2012 yet.

To fix this, let's just bump the snapshots.