Test suite: debug_log chutney events instead of puts:ing
puts() is intercepted by cucumber's pretty formatter: when run before
a scenario has started (i.e. initialize_chutney()) it works normally,
but when done during a step it is printed as blue text indented so it
"belongs" to the step. This looks bad.

These logging messages was mostly added for developers that are
monitoring a test suite run, since the wait can be long, but
developers will be looking at the debug log any way, so it is just
more suitable.