Test suite: disable the network for the VM by default
I.e. so the "the network is plugged" step is explicitly required in
order to have network access. This is useful for having that step as a
"hook" for stuff we want to run around the time we enable the
network (spoiler: it will be used to ensure that Chutney has
bootstrapped before we enable the network).