Test suite: be more careful about when we require Chutney to have bootstrapped
Now we ensure Chutney has bootstrapped before enabling the network in
the VM under testing. This makes it is impossible to write a test that
makes Tor connect over the network before Chutney is ready for
it. Before this I am sure there is some combination of steps that
would end up with Tor trying to connect without
wait_until_chutney_is_working() ever being run.

Also, we ensure Chutney is working after restoring from a snapshot in
order to avoid similar issues when running with --keep-snapshots.