onion-grater: craft correct empty answers for `getinfo circuit-status`
A '+' indicates that the answer is multi-line, a '-' that it is
single-line, which it will be if it is empty. See section 2.2 in
torspec/control-spec.txt.

My hope was that this would fix the issues I'm having while working on
the migration to Tor Browser 13 (refs: tails/tails#19478), where our
special-handling of `getinfo circuit-status` is breaking the circuit
view, but it didn't fix it.