Test suite: make sure browser URL bar is focused and ready for interaction
The "caller" of the "I open a new tab" step is now able to just
immediately type/paste the URL without risking it being lost. That is,
unless the dogtail-based method we use to determine if we focus the
correct element is good enough (previously we used a static 5 second
sleep).