Test suite: don't use translated string in Dogtail
The 'Search or enter address' string is translated in the localization
scenarios which would require us to extract the translations, which
isn't easy with Firefox' NIH translation system. So we attempt
something not relying on any translated strings.

Now I also realize that the above string is only used in Unsafe
Browesr, it is different in Tor Browser ("Search with DuckDuckGo ...")
so it was very inappropriate.