Test suite: don't focus the address bar in "open new tab" step
Now that all interactions with the address bar go through
`set_browser_url()` we only have to worry about it there.

While we're at it we also move the code that verifies that the address
bar is focused into `set_browser_url()`, to really make this
robust. Since that is separate from the dance we do to clear the
address bar and input the URL, we move the code about focusing outside
of the `try_for()`.