htpdate: drop leap.se, replacing it with gnu.org, replacing it with nextcloud.com
leap.se (friend) is sometimes reporting times that are 10+ minutes
incorrect so we need a better one. Since it is much easier to find
neutral sources than friends we promote gnu.org from neutral to
friend (which fits, they are privacy friendly, and fits with other
friends like eff.org). And then we add nextcloud.com to the neutral
pool as its replacement.

Refs: tails/tails#19923