Merge /lib/firmware → /usr/lib/firmware (refs: tails/tails#20075)
In recent Debian firmware packages (that we install from Sid) the
blobs have been migrated from /lib/firmware to /usr/lib/firmware which
works fine on a usr-merged system where /lib is a symlink to /usr/lib,
so both actually are the same directory. But in Tails 5.x, which isn't
usr-merged, this results in the firmwares being split among these
directories, and in the end only those in /lib/firmware are included
in the initramfs.

To fix this we implement a partial usr-merge: only /lib/firmware is
merged into /usr/lib/firmware, and then replaced with a symlink to
that directory.