Document a workaround for an issue when running Tails with virt-manager
When booting images built from feature/bookworm with my usual libvirt
configuration, tails-gdm-session fails to start, resulting boot to be
stuck on a black screen with the cursor blinking.

Here are some relevant error messages from the journal:

  gdm-shell-tails.desktop: LLVM ERROR: 64-bit code requested on a subtarget that doesn't support it!
  gdm-shell-tails.desktop: (EE) could not connect to wayland server
  gnome-session-binary: Unrecoverable failure in required component gdm-shell-tails.desktop

A workaround I found is to enable OpenGL in the Spice display settings.
We decided to document that in the Graphics issues section of
troubleshooting on the page for running Tails with virt-manager.