Test suite: fix scenario where Nautilus misbehaves with our showingOnly default (refs: tails/tails#19738)
When showingOnly = true it has a single child, the frame for the
selected tab, and that frame has no children so nothing is
accessible. This is not the case in Bullseye. I verified this directly
in a python session in Tails, so the fault is not in dogtail.rb's
integration into Ruby.

If more than two applications has this issue I might consider
refactoring so that the Dogtail::Application constructor can set the
default for showingOnly for each instance.