Tails Cloner: don't attempt to unmount the target device twice (refs: tails/tails#20139)
It leads to this failure when clicking Install:

    Unmounting mounted filesystems on "/dev/sda1"
    Unmounting mounted filesystems on "/dev/sda1"
    Could not get udisks object /dev/sda
    Tails installation failed!
    Could not get udisks object /dev/sda

i.e. it fails when attempting to unmount the already unmounted device.

This affects both Tails stable and feature/bookworm, it is just more
serious in the latter since we have enabled auto-mounting of removable
media (refs: tails/tails#15900).

Let's see if dropping the code that does the first unmounting works.