Test suite: use different method when filling storage devices until they are full
On feature/buster I often see the 'I fill the USB drive with a known
pattern' step fail with "execution expired (RemoteShell::Timeout)",
which means it took longer than 20 minutes. Although the previous
shell-loop-powered method is very slow it shouldn't be _that_
slow. But it truly writes at around 150 kB/s on my system... what?

It appears the problem is that we have enabled auto-mounting of
removable media (tails#15900), so in the 'I plug and mount a ... USB
drive with a ... filesystem' step the device is already mounted after
the plugging part, but then we manually mount it again. Why that makes
it slow I have no idea, but I'll fix this situation in a separate
commit.

I still think this new method is worthwhile to keep even though it
doesn't fix the real problem since it is faster (for whatever reason
it was also fast in the double-mount situation) and the code is a bit
more refactored.