Don't use stderr=subprocess.PIPE with subprocess.Popen
I assume `stderr=subprocess.PIPE` and `universal_newlines=True` was only set
here because the arguments were copied from the `subprocess.run` case
above. They make sense there - the `stderr=subprocess.PIPE` leads to any
`CalledProcessError` raised by `subprocess.run` having the `stderr` property
set and the `universal_newlines=True` results in that `stderr` property
being a string instead of bytes.

That's not the case with `subprocess.Popen`, which never raises a
`CalledProcessError`. In effect, the pipe that's being redirected to is
never used. With this commit, stderr is not redirected when using
`subprocess.Popen`.