Drop crappy doc and script to run rtorrent
It was announced as crappy right from the beginning and has not improved since.
It's an initscript written in a very old school manner, that makes shellcheck
legitimately unhappy. I don't think spreading this example is helpful.

I would bet the Internet has something better to offer nowadays,
a web search away.