Revert Polish changes made after Weblate upgrade
Polish is not activated in the production website, and we don't
currently have a reviewer for that language. Despite that, the upgrade
of Weblate done in Dec 14 2023 messed up with permissions and allowed
these changes to be committed to the repo.

Let's revert all changes made to Polish to their state from before the
problem happened.

refs tails#20132