Test suite: split out and fix "open the Report an Error launcher" step
Now it's doing something very different than the Tails Documentation
launcher does so they should be their own steps.

While we're at it we also simplify by dropping the retrying mechanism
for double-clicks: #12131 was an issue 7 years ago when we still used
Sikuli for click interactions. Now we use completely different
technology so let's see if it works better now, and otherwise see if
we can fix the root cause since we're in a better position to do so.