Avoid asp-post-apt hooks running forever in some cases
In the case that the user clicked "Create Persistent Storage" in the
notification but then doesn't actually create the Persistent Storage,
the asp-post-apt hook keeps running forever (or until eventually a
Persistent Storage is created).

To avoid wasting resources (especially memory) on those processes, we
now let them exit if the tps-frontend process exits without the
Persistent Storage being created.