Tails Cloner: unmount all partitions related to the target when needed
No partition (e.g. "siblings") on the target device can be mounted (or
unlocked, if a LUKS device) while partitioning or installing the boot
loader, so we have to deal with that.

This commit basically rewrites unmount_device() completely to make it
act on up-to-date information (instead of possibly outdated and often
completely incorrect information stored in `data['mount']` and
`data['mounted_partitions']`) and also deal with locking of
LUKS-encrypted volumes. The new boolean flag `unmount_all` was added
to deal with _other_ partitions on the target device.

Notably it is now also possible to install Tails to an encrytped
target, which previously wasn't possible. I'm not sure why this was
ever the case, we made it so in a commit with a useless commit message
that doesn't explain why.

Refs: #20149