Revert "Avoid asp-post-apt hooks running forever in some cases"
This reverts commit a609edbd23b79c5ed95a75e851327e8960dad078.

I've seen a case when live-additional-software.conf was not created: we're stuck
at "Waiting for Persistent Storage to be ready", while tps-frontend is left
running, and tails-additional-software-post-apt-hook.service has
terminated successfully.

I understand that's because a609edbd introduced a race condition: the way we
spawn tps-frontend, systemd-run "will start the service asynchronously in the
background and return after the command has begun execution" — the command being
run-with-env. So we have no guarantee that a tps-frontend process is already by
the time spawn_tps_frontend() returns.